+++ src/react/src/components/FilterSidebarExtendedSearch.jsx
@@ -21,7 +21,6 @@ import {
     fetchContributorTypeOptions,
     fetchFacilityProcessingTypeOptions,
     fetchNumberOfWorkersOptions,
-    fetchParentCompanyOptions,
 } from '../actions/filterOptions';
 
 import {
@@ -32,7 +31,6 @@ import {
     facilityProcessingTypeOptionsPropType,
     productTypeOptionsPropType,
     numberOfWorkerOptionsPropType,
-    parentCompanyOptionsPropType,
 } from '../util/propTypes';
 
 import {
@@ -46,6 +44,7 @@ const FACILITY_TYPE = 'FACILITY_TYPE';
 const PROCESSING_TYPE = 'PROCESSING_TYPE';
 const PRODUCT_TYPE = 'PRODUCT_TYPE';
 const NUMBER_OF_WORKERS = 'NUMBER_OF_WORKERS';
+const customSelectComponents = { DropdownIndicator: null };
 
 const mapFacilityTypeOptions = (fPTypes, pTypes) => {
     let fTypes = [];
@@ -88,7 +87,6 @@ const isExtendedFieldForThisContributor = (field, extendedFields) =>
 
 function FilterSidebarExtendedSearch({
     contributorTypeOptions,
-    parentCompanyOptions,
     facilityProcessingTypeOptions,
     numberOfWorkersOptions,
     contributorTypes,
@@ -108,7 +106,6 @@ function FilterSidebarExtendedSearch({
     embed,
     embedExtendedFields,
     fetchContributorTypes,
-    fetchParentCompanies,
     fetchFacilityProcessingType,
     fetchNumberOfWorkers,
     isSideBarSearch,
@@ -118,12 +115,6 @@ function FilterSidebarExtendedSearch({
             fetchContributorTypes();
         }
     }, [contributorTypeOptions, fetchContributorTypes]);
-
-    useEffect(() => {
-        if (!parentCompanyOptions) {
-            fetchParentCompanies();
-        }
-    }, [parentCompanyOptions, fetchParentCompanies]);
 
     useEffect(() => {
         if (!facilityProcessingTypeOptions) {
@@ -177,12 +168,14 @@ function FilterSidebarExtendedSearch({
                         creatable
                         label="Parent Company"
                         name={PARENT_COMPANY}
-                        options={parentCompanyOptions || []}
                         value={parentCompany}
                         onChange={updateParentCompany}
-                        disabled={fetchingExtendedOptions || fetchingFacilities}
                         placeholder="e.g. ABC Textiles Limited"
                         isSideBarSearch={isSideBarSearch}
+                        aria-label="Parent
+                    company"
+                        components={customSelectComponents}
+                        disabled={fetchingFacilities}
                     />
                 </div>
             </ShowOnly>
@@ -283,14 +276,12 @@ function FilterSidebarExtendedSearch({
 
 FilterSidebarExtendedSearch.defaultProps = {
     contributorTypeOptions: null,
-    parentCompanyOptions: null,
     facilityProcessingTypeOptions: null,
     numberOfWorkersOptions: null,
 };
 
 FilterSidebarExtendedSearch.propTypes = {
     contributorTypeOptions: contributorTypeOptionsPropType,
-    parentCompanyOptions: parentCompanyOptionsPropType,
     facilityProcessingTypeOptions: facilityProcessingTypeOptionsPropType,
     numberOfWorkersOptions: numberOfWorkerOptionsPropType,
     updateContributorType: func.isRequired,
@@ -309,10 +300,6 @@ function mapStateToProps({
         contributorTypes: {
             data: contributorTypeOptions,
             fetching: fetchingContributorTypes,
-        },
-        parentCompanies: {
-            data: parentCompanyOptions,
-            fetching: fetchingParentCompanies,
         },
         facilityProcessingType: {
             data: facilityProcessingTypeOptions,
@@ -339,7 +326,6 @@ function mapStateToProps({
 }) {
     return {
         contributorTypeOptions,
-        parentCompanyOptions,
         facilityProcessingTypeOptions,
         numberOfWorkersOptions,
         contributorTypes,
@@ -354,8 +340,7 @@ function mapStateToProps({
         fetchingExtendedOptions:
             fetchingContributorTypes ||
             fetchingFacilityProcessingType ||
-            fetchingNumberofWorkers ||
-            fetchingParentCompanies,
+            fetchingNumberofWorkers,
         embed: !!embed,
         embedExtendedFields: config.extended_fields,
     };
@@ -372,7 +357,6 @@ function mapDispatchToProps(dispatch) {
         updateNativeLanguageName: e =>
             dispatch(updateNativeLanguageNameFilter(getValueFromEvent(e))),
         fetchContributorTypes: () => dispatch(fetchContributorTypeOptions()),
-        fetchParentCompanies: () => dispatch(fetchParentCompanyOptions()),
         fetchFacilityProcessingType: () =>
             dispatch(fetchFacilityProcessingTypeOptions()),
         fetchNumberOfWorkers: () => dispatch(fetchNumberOfWorkersOptions()),